Separating out Rubee Module classes to separate files.
